Refactored solr request converter code
