+++ src/services/ngram-trends/logic/solrQuery.js
@@ -1,0 +1,9 @@
+const { filtersToQueryAndVariables } = require('../../../util/solr');
+
+function ngramTrendsRequestToSolrQuery(ngrams, filters, facets) {
+  const { query, variables } = filtersToQueryAndVariables(filters);
+}
+
+module.exports = {
+  ngramTrendsRequestToSolrQuery,
+};
